+++ src/shuriken/lazy.clj
@@ -1,0 +1,4 @@
+(ns shuriken.lazy)
+
+(defn deep-doall [m]
+  (doall (clojure.walk/postwalk identity m)))